Fix for i18n >= 1.2.0
Version 1.2.0 of the i18n gem changed the way the regexp for
interpolation arguments is stored, and added an official way to include
additional patterns (see ruby-i18n/i18n@196d144).

Although this is a backward-incompatible change (i.e. will not work with
i18n < 1.2.0), it does not require a new major version of the
dependency, so it is not considered a breaking change.